ROManifolds.ParamFESpaces

ROManifolds.ParamFESpaces.MultiFieldParamFEFunction — Type

struct MultiFieldParamFEFunction{T<:MultiFieldCellField} <: ParamFEFunction
  single_fe_functions::Vector{<:SingleFieldParamFEFunction}
  free_values::AbstractArray
  fe_space::MultiFieldFESpace
  multi_cell_field::T
end

ROManifolds.ParamFESpaces.MultiFieldParamFESpace — Type

const MultiFieldParamFESpace = MultiFieldFESpace

ROManifolds.ParamFESpaces.ParamFEFunction — Type

abstract type ParamFEFunction <: FEFunction end

Parametric extension of a FEFunction in Gridap. Subtypes:

SingleFieldParamFEFunction

MultiFieldParamFEFunction

ROManifolds.ParamFESpaces.SingleFieldParamFEFunction — Type

struct SingleFieldParamFEFunction{T<:CellField} <: ParamFEFunction
  cell_field::T
  cell_dof_values::AbstractArray{<:AbstractParamVector{<:Number}}
  free_values::AbstractParamVector{<:Number}
  dirichlet_values::AbstractParamVector{<:Number}
  fe_space::SingleFieldFESpace
end

ROManifolds.ParamFESpaces.SingleFieldParamFESpace — Type

abstract type SingleFieldParamFESpace{S} <: SingleFieldFESpace end

Parametric extension of a SingleFieldFESpace in Gridap. The FE spaces inhereting are (trial) spaces on which we can easily define a ParamFEFunction.

Subtypes:

TrivialParamFESpace

TrialParamFESpace

ROManifolds.ParamFESpaces.TrialParamFESpace — Type

TrialParamFESpace{S} <: SingleFieldParamFESpace{S}

Trial FE space equipped with parametric dirichlet values

ROManifolds.ParamFESpaces.TrivialParamFESpace — Type

TrivialParamFESpace{S} <: SingleFieldParamFESpace{S}

Wrapper for non-parametric FE spaces that we wish assumed a parametric length

ROManifolds.ParamDataStructures.parameterize — Method

parameterize(a::SparseMatrixAssembler,r::AbstractRealization) -> SparseMatrixAssembler

Returns an assembler that also stores the parametric length of r. This function is to be used to assemble parametric residuals and Jacobians. The assembly routines follow the same pipeline as in Gridap

ROManifolds.ParamFESpaces.HomogeneousTrialParamFESpace — Method

HomogeneousTrialParamFESpace(U::SingleFieldFESpace,plength::Int) -> TrialParamFESpace

Returns a TrialParamFESpace equipped with homogeneous parametric dirichlet values

ROManifolds.ParamFESpaces.TrialParamFESpace! — Method

TrialParamFESpace!(dir_values::AbstractParamVector,space::SingleFieldFESpace,objects
  ) -> TrialParamFESpace

Allows do-block syntax for the construction of a TrialParamFESpace

ROManifolds.ParamFESpaces.get_vector_type2 — Method

get_vector_type2(f::SingleFieldParamFESpace) -> Type

Returns the vector type of the underlying un-parametric FESpace contained in f

ROManifolds.ParamFESpaces.param_zero_dirichlet_values — Method

param_zero_dirichlet_values(f::FESpace) -> AbstractParamVector

Parametric version of the function zero_dirichlet_values in Gridap

ROManifolds.ParamFESpaces.param_zero_free_values — Method

param_zero_free_values(f::FESpace) -> AbstractParamVector

Parametric version of the function zero_free_values in Gridap