ROM.ParamODEs

ROM.ParamODEs.JointTransientParamFEOperator — Type

const JointTransientParamFEOperator{O<:ODEParamOperatorType} = TransientParamFEOperator{O,JointDomains}

ROM.ParamODEs.LinearNonlinearParamODE — Type

struct LinearNonlinearParamODE <: ODEParamOperatorType end

ROM.ParamODEs.LinearNonlinearParamOpFromTFEOp — Type

ROM.ParamODEs.LinearNonlinearTransientParamFEOperator — Type

struct LinearNonlinearTransientParamFEOperator{T} <: TransientParamFEOperator{LinearNonlinearParamODE,T}
  op_linear::TransientParamFEOperator{LinearParamODE,T}
  op_nonlinear::TransientParamFEOperator{NonlinearParamODE,T}
end

Interface to accommodate the separation of terms depending on their linearity in a nonlinear problem. This allows to build and store once and for all linear residuals/Jacobians, and in the Newton-like iterations only evaluate and assemble only the nonlinear components

ROM.ParamODEs.LinearParamODE — Type

struct LinearParamODE <: ODEParamOperatorType end

ROM.ParamODEs.NonlinearParamODE — Type

struct NonlinearParamODE <: ODEParamOperatorType end

ROM.ParamODEs.ODEParamOpFromTFEOp — Type

struct ODEOpFromTFEOp <: ODEOperator end

Wrapper that transforms a TransientParamFEOperator into an ODEParamOperator

ROM.ParamODEs.ODEParamOperator — Type

abstract type ODEParamOperator{T<:ODEParamOperatorType,T<:TriangulationStyle} <: ParamOperator{O,T} end

Transient extension of the type ParamOperator.

ROM.ParamODEs.ODEParamOperatorType — Type

abstract type ODEParamOperatorType <: UnEvalOperatorType end

Parametric extension of the type ODEOperatorType in Gridap.

Subtypes:

NonlinearParamODE

LinearParamODE

LinearNonlinearParamODE

ROM.ParamODEs.ODEParamSolution — Type

struct ODEParamSolution{V} <: ODESolution
  solver::ODESolver
  odeop::ODEParamOperator
  r::TransientRealization
  us0::Tuple{Vararg{V}}
  tracker::CostTracker
end

Wrapper for the evolution of a differential problem represented by the field odeop, and solved by means of the ode solver solver. It represents the parametric extension of the type ODESolution in Gridap

ROM.ParamODEs.ParamStageOperator — Type

struct ParamStageOperator{O} <: StageOperator
  op::ODEParamOperator{O}
  cache::AbstractParamCache
  r::TransientRealization
  us::Function
  ws::Tuple{Vararg{Real}}
end

(Nonlinear) stage operator to solve a parametric ODE with a time marching scheme

ROM.ParamODEs.SplitTransientParamFEOperator — Type

const SplitTransientParamFEOperator{O<:ODEParamOperatorType} = TransientParamFEOperator{O,SplitDomains}

ROM.ParamODEs.TransientMultiFieldParamFESpace — Type

const TransientMultiFieldParamFESpace = MultiFieldFESpace

ROM.ParamODEs.TransientParamFEOpFromWeakForm — Type

struct TransientParamFEOpFromWeakForm{T} <: TransientParamFEOperator{NonlinearParamODE,T}
  res::Function
  jacs::Tuple{Vararg{Function}}
  tpspace::TransientParamSpace
  assem::Assembler
  trial::FESpace
  test::FESpace
  domains::FEDomains
  order::Integer
end

Instance of TransientParamFEOperator, to be used when the transient problem is nonlinear

ROM.ParamODEs.TransientParamFEOperator — Type

abstract type TransientParamFEOperator{T<:ODEParamOperatorType} <: ParamFEOperator{T} end

Parametric extension of a TransientFEOperator in Gridap. Compared to a standard TransientFEOperator, there are the following novelties:

a TransientParamSpace is provided, so that parametric realizations can be extracted directly from the TransientParamFEOperator

a function representing a norm matrix is provided, so that errors in the desired norm can be automatically computed

Subtypes:

TransientParamFEOpFromWeakForm

TransientParamLinearFEOpFromWeakForm

LinearNonlinearTransientParamFEOperator

ROM.ParamODEs.TransientParamFESolution — Type

struct TransientParamFESolution{V} <: TransientFESolution
  odesol::ODEParamSolution{V}
  trial
end

Wrapper around a TransientParamFEOperator and ODESolver that represents the parametric solution at a set of time steps. It is an iterator that computes the solution at each time step in a lazy fashion when accessing the solution.

ROM.ParamODEs.TransientParamLinearFEOpFromWeakForm — Type

struct TransientParamLinearFEOpFromWeakForm{T} <: TransientParamFEOperator{LinearParamODE,T}
  res::Function
  jacs::Tuple{Vararg{Function}}
  constant_forms::Tuple{Vararg{Bool}}
  tpspace::TransientParamSpace
  assem::Assembler
  trial::FESpace
  test::FESpace
  domains::FEDomains
  order::Integer
end

Instance of TransientParamFEOperator, to be used when the transient problem is linear

ROM.ParamODEs.TransientTrialParamFESpace — Type

const TransientTrialParamFESpace = UnEvalTrialFESpace

ROM.ParamODEs.TransientParamLinearFEOperator — Method

TransientParamLinearFEOperator(forms::Tuple{Vararg{Function}},res::Function, tpspace,trial,test;kwargs...) -> TransientParamLinearFEOpFromWeakForm{TriangulationStyle}

Returns a linear parametric FE operator

ROM.ParamODEs.collect_initial_values — Method

collect_initial_values(sol::ODEParamSolution) -> AbstractParamVector

Fetches the initial values of a ODEParamSolution sol