ROManifolds.ParamSteady

ROManifolds.ParamSteady.FEDomains — Type

struct FEDomains{A,B}
  domains_res::A
  domains_jac::B
end

Fields:

domains_res: triangulations relative to the residual (nothing by default)

domains_jac: triangulations relative to the Jacobian (nothing by default)

ROManifolds.ParamSteady.JointDomains — Type

struct JointDomains <: TriangulationStyle end

Trait for a FE operator indicating that residuals/Jacobiansin this operator should be computed summing the contributions relative to each triangulation as occurs in Gridap

ROManifolds.ParamSteady.JointParamFEOperator — Type

const JointParamFEOperator{O<:UnEvalOperatorType} = ParamFEOperator{O,JointDomains}

ROManifolds.ParamSteady.JointParamOpFromFEOp — Type

const JointParamOpFromFEOp{O<:UnEvalOperatorType} = ParamOpFromFEOp{O,JointDomains}

ROManifolds.ParamSteady.JointParamOperator — Type

const JointParamOperator{O<:UnEvalOperatorType} = ParamOperator{O,JointDomains}

ROManifolds.ParamSteady.LinearNonlinearParamEq — Type

struct LinearNonlinearParamEq <: UnEvalOperatorType end

ROManifolds.ParamSteady.LinearNonlinearParamFEOperator — Type

struct LinearNonlinearParamFEOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: ParamFEOperator{O,T}
  op_linear::ParamFEOperator
  op_nonlinear::ParamFEOperator
end

Interface to accommodate the separation of terms depending on their linearity in a nonlinear problem. This allows to build and store once and for all linear residuals/Jacobians, and in the Newton-like iterations only evaluate and assemble only the nonlinear components

ROManifolds.ParamSteady.LinearParamEq — Type

struct LinearParamEq <: UnEvalOperatorType end

ROManifolds.ParamSteady.NonlinearParamEq — Type

struct NonlinearParamEq <: UnEvalOperatorType end

ROManifolds.ParamSteady.ParamFEOpFromWeakForm — Type

struct ParamFEOpFromWeakForm{O<:UnEvalOperatorType,T<:TriangulationStyle} <: ParamFEOperator{O,T}
  res::Function
  jac::Function
  pspace::ParamSpace
  assem::Assembler
  trial::FESpace
  test::FESpace
  domains::FEDomains
end

Most standard instance of a parametric FE operator

ROManifolds.ParamSteady.ParamFEOperator — Type

abstract type ParamFEOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: FEOperator end

Parametric extension of a FEOperator in Gridap. Compared to a standard FEOperator, there are the following novelties:

a ParamSpace is provided, so that parametric realizations can be extracted directly from the ParamFEOperator

a function representing a norm matrix is provided, so that errors in the desired norm can be automatically computed

Subtypes:

ParamFEOpFromWeakForm

LinearNonlinearParamFEOperator

TransientParamFEOperator

ROManifolds.ParamSteady.ParamOpFromFEOp — Type

struct ParamOpFromFEOp{O<:UnEvalOperatorType,T<:TriangulationStyle} <: ParamOperator{O,T}
  op::ParamFEOperator{O,T}
end

Wrapper that transforms a ParamFEOperator into an ParamOperator

ROManifolds.ParamSteady.ParamOperator — Type

abstract type ParamOperator{O<:UnEvalOperatorType,T<:TriangulationStyle} <: NonlinearParamOperator end

Type representing algebraic operators when solving parametric differential problems

ROManifolds.ParamSteady.ParamTrialFESpace — Type

const ParamTrialFESpace = UnEvalTrialFESpace

ROManifolds.ParamSteady.SplitDomains — Type

struct SplitDomains <: TriangulationStyle end

Trait for a FE operator indicating that residuals/Jacobiansin this operator should be computed keeping the contributions relative to each triangulation separate

ROManifolds.ParamSteady.SplitParamFEOperator — Type

const SplitParamFEOperator{O<:UnEvalOperatorType} = ParamFEOperator{O,SplitDomains}

ROManifolds.ParamSteady.SplitParamOpFromFEOp — Type

const SplitParamOpFromFEOp{O<:UnEvalOperatorType} = ParamOpFromFEOp{O,SplitDomains}

ROManifolds.ParamSteady.SplitParamOperator — Type

const SplitParamOperator{O<:UnEvalOperatorType} = ParamOperator{O,SplitDomains}

ROManifolds.ParamSteady.TriangulationStyle — Type

abstract type TriangulationStyle <: GridapType end

Subtypes:

JointDomains

SplitDomains

ROManifolds.ParamSteady.UnEvalOperatorType — Type

abstract type UnEvalOperatorType <: GridapType end

Type representing operators that are not evaluated yet. This may include operators representing transient problems (although the implementation in Gridap differs), parametric problems, and a combination thereof. Could become a supertype of ODEOperatorType in Gridap

ROManifolds.ParamSteady.UnEvalTrialFESpace — Type

struct UnEvalTrialFESpace{A,B} <: SingleFieldFESpace
  space::A
  space0::B
  dirichlet::Union{Function,AbstractVector{<:Function}}
end

Struct representing trial FE spaces that are not evaluated yet. This may include FE spaces representing transient problems (although the implementation in Gridap, called TransientTrialFESpace, does not fall into this category), parametric problems, and transient-parametric problems.

ROManifolds.ParamSteady.LinearParamFEOperator — Method

LinearParamFEOperator(res::Function,jac::Function,pspace,trial,test,args...
  ) -> ParamFEOperator{LinearParamEq,TriangulationStyle}

Returns a linear parametric FE operator

ROManifolds.ParamSteady.collect_cell_matrix_for_trian — Method

function collect_cell_matrix_for_trian(
  trial::FESpace,
  test::FESpace,
  a::DomainContribution,
  strian::Triangulation
  ) -> Tuple{Vector{<:Any},Vector{<:Any},Vector{<:Any}}

Computes the cell-wise data needed to assemble a global sparse matrix for a given input triangulation strian

ROManifolds.ParamSteady.collect_cell_vector_for_trian — Method

function collect_cell_vector_for_trian(
  test::FESpace,
  a::DomainContribution,
  strian::Triangulation
  ) -> Tuple{Vector{<:Any},Vector{<:Any}}

Computes the cell-wise data needed to assemble a global vector for a given input triangulation strian

ROManifolds.ParamSteady.get_dof_map_at_domains — Method

get_dof_map_at_domains(op::ParamFEOperator) -> Contribution

Returns the residual dof map restricted to every residual triangulation

ROManifolds.ParamSteady.get_fe_operator — Method

get_fe_operator(op::ParamOperator) -> ParamFEOperator

Fetches the underlying FE operator of an algebraic operator op

ROManifolds.ParamSteady.get_param_space — Method

get_param_space(feop::ParamFEOperator) -> ParamSpace
get_param_space(feop::TransientParamFEOperator) -> TransientParamSpace

Returns the space of parameters contained in feop

ROManifolds.ParamSteady.get_sparse_dof_map_at_domains — Method

get_sparse_dof_map_at_domains(op::ParamFEOperator) -> Contribution

Returns the Jacobian dof map restricted to every Jacobian triangulation

ROManifolds.ParamSteady.join_operators — Method

join_operators(op::LinearNonlinearParamFEOperator) -> ParamFEOperator

Joins the linear/nonlinear parts of the operator and returns the resulting operator

ROManifolds.Utils.change_domains — Method

change_domains(op::ParamFEOperator,args...) -> ParamFEOperator

Changes the triangulations for residual/jacobian; the resulting operator will have the same TriangulationStyle trait as the one of the input

ROManifolds.Utils.set_domains — Method

set_domains(op::ParamFEOperator,args...) -> JointParamFEOperator

Fixes the triangulations for residual/jacobian; the resulting operator will have the trait JointDomains activated