ROM.RBTransient

ROM.RBTransient.GenericTransientRBOperator — Type

struct GenericTransientRBOperator{O} <: TransientRBOperator{O}
  op::ODEParamOperator{O}
  trial::RBSpace
  test::RBSpace
  lhs::TupOfAffineContribution
  rhs::AffineContribution
end

Transient counterpart of a GenericRBOperator] used in steady problems. Fields:

op: underlying high dimensional FE operator

trial: reduced trial space

test: reduced trial space

lhs: hyper-reduced left hand side

rhs: hyper-reduced right hand side

The major difference with respect to the steady setting is that the lhs is a n-tuple of AffineContribution, where n is the maximum order of the time derivatives

ROM.RBTransient.LinearNonlinearTransientRBOperator — Type

struct LinearNonlinearTransientRBOperator <: TransientRBOperator{LinearNonlinearParamODE}
  op_linear::GenericTransientRBOperator{LinearParamODE}
  op_nonlinear::GenericTransientRBOperator{NonlinearParamODE}
end

Extends the concept of GenericTransientRBOperator to accommodate the linear/nonlinear splitting of terms in nonlinear applications

ROM.RBTransient.TransientIntegrationDomain — Type

struct TransientIntegrationDomain <: AbstractIntegrationDomain{AbstractVector{Int}}
  indices_space::IntegrationDomain
  indices_time::IntegrationDomain
end

Integration domain for a projection operator in a transient context

ROM.RBTransient.TransientMDEIMReduction — Type

struct TransientMDEIMReduction{A,R<:Reduction{A,EuclideanNorm}} <: AbstractMDEIMReduction{A}
  reduction::R
  combine::Function
end

MDEIM struct employed in transient problems. The field combine is a function used to group the reductions relative to the various Jacobians(in general, more than one in transient problems) in a smart way. We consider, for example, the ODE

du/dt - νΔu = f in Ω × [0,T]

subject to initial/boundary conditions. Upon applying a FE discretization in space, and a θ-method in time, one gets the space-time system Aθ * uθ = fθ, where

Aθ = [ A₁ + M / (θ*Δt) 0 0 ⋯ 0 0 - M / (θ*Δt) A₂ + M / (θ*Δt) 0 ⋯ 0 0 0 - M / (θ*Δt) A₃ + M / (θ*Δt) 0 0 ⋱ ⋱ ⋱ ⋯ ⋯ ⋯ 0 0 0 ⋯ - M / (θ*Δt) Aₙ + M / (θ*Δt) ] = tridiag(- M/(θ*Δt), Aₖ + M/(θ*Δt), 0) uθ = [(1-θ)u₀ + θu₁, (1-θ)u₁ + θu₂ ⋯ (1-θ)uₙ₋₁ + θuₙ] fθ = [f₁, f₂ ⋯ fₙ]

where Aₖ = A(tₖ₋₁ + θ*Δt) and fₖ = f(tₖ₋₁ + θ*Δt).

Note: instead of multiplying Aθ by uθ, we multiply Ãθ by u, where

Ãθ = tridiag((1-θ)Aₖ₋₁ - M/Δt, θAₖ + M/Δt, 0) u = [u₁, u₂ ⋯ uₙ]

We now denote with Φ, Ψ the spatial and temporal basis obtained by reducing the snapshots associated to the state variable u. The Galerkin projection of the space-time system is equal to Âθ * û = f̂θ, where û is the unknown, and

Âθ = ∑ₖⁿ⁻¹ ( (1-θ)*ΦᵀAₖΦ - ΦᵀMΦ / Δt ) ⊗ Ψ[k-1,:]ᵀΨ[k,:] + ∑ₖⁿ ( θ*ΦᵀAₖΦ + ΦᵀMΦ / Δt ) ⊗ Ψ[k,:]ᵀΨ[k,:] f̂θ = ∑ₖⁿ Φᵀfₖ ⊗ Ψ[k,:]

We notice that the expression of Âθ can be written in a more general form as

Âθ = combine_A(A shift back, A) + combine_M(M shift back, M)

where combineA and combineM are two function specific to A and M:

combine_A(x,y) = θ*y + (1-θ)*y combine_M(x,y) = (x - y) / Δt

The same can be said of any time marching scheme. This is the meaning of the function combine. Note that for a time marching with p interpolation points (e.g. for θ method, p = 2) the combine functions will have to accept p arguments.

ROM.RBTransient.TransientProjection — Type

struct TransientProjection <: Projection
  projection_space::Projection
  projection_time::Projection
end

Projection operator for transient problems, containing a spatial projection and a temporal one. The space-time projection operator is equal to

projection_time ⊗ projection_space

which, for efficiency reasons, is never explicitly computed

ROM.RBTransient.TransientRBOperator — Type

abstract type TransientRBOperator{O} <: ODEParamOperator{O,SplitDomains} end

Type representing reduced algebraic operators used within a reduced order modelling framework in transient applications. A TransientRBOperator should contain the following information:

a reduced test and trial space, computed according to reduced_spaces

a hyper-reduced residual and jacobian, computed according to reduced_weak_form

Subtypes:

GenericTransientRBOperator

LinearNonlinearTransientRBOperator

ROM.RBTransient.TransientReduction — Type

struct TransientReduction{A,B,RS<:Reduction{A,B},RT<:Reduction{A,EuclideanNorm}} <: Reduction{A,B}
  reduction_space::RS
  reduction_time::RT
end

Wrapper for reduction methods in transient problems. The fields reduction_space and reduction_time respectively represent the spatial reduction method, and the temporal one

ROM.RBTransient.time_enrichment — Method

time_enrichment(red::SupremizerReduction,a_primal::Projection,basis_dual) -> AbstractMatrix

Temporal supremizer enrichment. (Approximate) Procedure:

for every b_dual ∈ Col(basis_dual)

compute Φ_primal_dual = get_basis(a_primal)'*get_basis(b_dual)

compute v = kernel(Φ_primal_dual)

compute v′ = orth_complement(v,a_primal)

enrich a_primal = [a_primal,v′]