ROManifolds.Utils

ROManifolds.Utils.ArrayContribution — Type

struct ArrayContribution{T,N,V,K} <: Contribution
  values::V
  trians::K
end

Contribution whose field values are AbstractArrays

ROManifolds.Utils.Contribution — Type

abstract type Contribution end

Represents quantitites whose values vary upon a triangulation. The values can be accessed by indexing a the corresponding triangulation. See DomainContribution in Gridap for more details.

ROManifolds.Utils.MatrixContribution — Type

const MatrixContribution{T,V,K} = ArrayContribution{T,2,V,K}

ROManifolds.Utils.PartialDerivative — Type

struct PartialDerivative{N} <: Function end

Gridap Map that implements a partial derivative

ROManifolds.Utils.TupOfArrayContribution — Type

const TupOfArrayContribution{T} = Tuple{Vararg{ArrayContribution{T}}}

Specifically allows to deal with tuples of Jacobians in unsteady settings

ROManifolds.Utils.VectorContribution — Type

const VectorContribution{T,V,K} = ArrayContribution{T,1,V,K}

ROManifolds.Utils.∂₁ — Type

const ∂₁ = PartialDerivative{1}

ROManifolds.Utils.∂₂ — Type

const ∂₂ = PartialDerivative{2}

ROManifolds.Utils.∂₃ — Type

const ∂₃ = PartialDerivative{3}

ROManifolds.Utils.compute_error — Method

compute_error(
  sol::AbstractArray{T,N},
  sol_approx::AbstractArray{T,N},
  args...
  ) where {T,N} -> Number

Computes the error between sol and sol_approx, by default in the Euclidean norm. A different norm (usually represented by a sparse matrix) can be provided as an argument.

ROManifolds.Utils.compute_relative_error — Method

compute_relative_error(
  sol::AbstractArray{T,N},
  sol_approx::AbstractArray{T,N},
  args...
  ) where {T,N} -> Number

Computes the relative error between sol and sol_approx, by default in the Euclidean norm. A different norm (usually represented by a sparse matrix) can be provided as an argument.

ROManifolds.Utils.compute_speedup — Method

compute_speedup(t1::CostTracker,t2::CostTracker) -> Speedup

Computes the speedup the tracker t2 achieves with respect to t1, in time and in memory footprint

ROManifolds.Utils.contribution — Method

contribution(f,trians) -> Contribution

Constructor of a Contribution that allows do-block syntax. f is a function such that values[i] = f(trians[i]) for i... This constructor first builds the tuple of values, then builds the Contribution object from values and trians

ROManifolds.Utils.find_closest_view — Method

find_closest_view(
  tparents::Tuple{Vararg{Triangulation}},
  tchild::Triangulation
  ) -> (Integer, Triangulation)

Finds the approximate parent of tchild; it returns the parent's index and its view in the same indices as tchild (which should be a triangulation view)

ROManifolds.Utils.get_parent — Method

get_parent(t::Triangulation) -> Triangulation

When t is a triangulation view, returns its parent; throws an error when t is not a triangulation view

ROManifolds.Utils.is_included — Method

is_included(tchild::Triangulation,tparent::Triangulation) -> Bool

Returns true if tchild is a triangulation included in tparent, false otherwise. This condition is not as strong as is_parent

ROManifolds.Utils.is_parent — Method

is_parent(tparent::Triangulation,tchild::Triangulation) -> Bool

Returns true if tchild is a triangulation view of tparent, false otherwise

ROManifolds.Utils.merge_triangulations — Method

merge_triangulations(trians::AbstractVector{<:Triangulation}) -> Triangulation

Given a tuple of triangulation views trians, returns the triangulation view on the union of the viewed cells. In other words, the minimum common integration domain is found

ROManifolds.Utils.order_domains — Method

order_domains(
  tparents::Tuple{Vararg{Triangulation}},
  tchildren::Tuple{Vararg{Triangulation}}
  ) -> Tuple{Vararg{Triangulation}}

Orders the triangulation children in the same way as the triangulation parents